tools/meeting_assistant/requirements.txt
paho-mqtt>=1.6.1,<2.0
pygetwindow>=0.0.9
keyboard>=0.13.5
soundcard>=0.4.2
numpy>=1.24.0
groq>=0.4.0
pyinstaller>=6.0
Pillow>=10.0
pystray>=0.19
